feat(issuer-api): add issuer module options BREAKING CHANGE: issuer-api AppModule was renamed to IssuerModule. Also, to use it, you need to call `IssuerModule.register()`
