+++ packages/traceability/issuer-api/src/const.ts
@@ -1,0 +1,1 @@
+export const ISSUER_MODULE_OPTIONS_TOKEN = Symbol.for('ISSUER_MODULE_OPTIONS_TOKEN');
