+++ packages/traceability/issuer-api/src/types/index.ts
@@ -10,3 +10,7 @@ export class CertificateEvent {
     @ApiProperty()
     values?: any;
 }
+
+export interface IssuerModuleOptions {
+    enableTransactionLogging: boolean;
+}
